patch to keep track of offset time in metadata and send it out to the player.
This allows us to present accurate presentation time for a given clip that is it-self a segment of a longer clip. (for example a segment downloaded from archive.org) This applies to oggz_chop segments being played with the mv_embed javascript player where the controls are based in html. Having the time offset is necessary to build a playhead from the offset timestamps given in callbacks during playback.

(I would have submitted a bug and patch to bugzilla but its giving me trouble)