Change the width of the div surrouding mediaplayers into min-width
The native a/v players of FireFox and Chrome, have a minimum width that is
rather large (220px). In many cases Wikipedia thumbs are requested at a size 
that is smaller than that. With a set width, the elements than overflow out
of fixed width divs. Using min-width instead allows the surrounding div
to resize to fit its contents.

Fixes bug 22538